pointerstuff

+++ ciphers.go
@@ -389,11 +389,11 @@ func newSeal(c *Cipher, k []byte) (*sealCipherCtx, error) {
 	symKey := make([]byte, pkeySize)
 
 	IV := make([]byte, ctx.IVSize())
-	var symKeySize int
+	var symKeySize C.int
 
 	if 0 == C.EVP_SealInit(ctx.ctx, c.ptr,
 		(**C.uchar)(unsafe.Pointer((*C.uchar)(&symKey[0]))),
-		(*C.int)(&C.int(symKeySize)),
+		(*C.int)(&symKeySize),
 		(*C.uchar)(&IV[0]),
 		(**C.EVP_PKEY)(unsafe.Pointer(key.evpPKey())),
 		C.int(1)) {
